requirements.txt
aiohttp==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
elasticsearch==7.10.1
googlesearch-python==2020.0.2
idna==2.10
multidict==4.7.6
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.9
yarl==1.5.1
